Use a common ShaderId class. Start moving shader ID code to GPU/Common
