+++ GPU/Directx9/PixelShaderGeneratorDX9.h
@@ -19,34 +19,14 @@
 
 #include "Globals.h"
 
+#include "GPU/Common/ShaderId.h"
+
 // TODO: Bench both ways. Result may be different on old vs new hardware though..
 // #define DX9_USE_HW_ALPHA_TEST 1
 
 namespace DX9 {
 
-struct FragmentShaderIDDX9 {
-	FragmentShaderIDDX9() {clear();}
-	void clear() {d[0] = 0xFFFFFFFF; d[1] = 0xFFFFFFFF;}
-	u32 d[2];
-	bool operator < (const FragmentShaderIDDX9 &other) const {
-		for (size_t i = 0; i < sizeof(d) / sizeof(u32); i++) {
-			if (d[i] < other.d[i])
-				return true;
-			if (d[i] > other.d[i])
-				return false;
-		}
-		return false;
-	}
-	bool operator == (const FragmentShaderIDDX9 &other) const {
-		for (size_t i = 0; i < sizeof(d) / sizeof(u32); i++) {
-			if (d[i] != other.d[i])
-				return false;
-		}
-		return true;
-	}
-};
-
-void ComputeFragmentShaderIDDX9(FragmentShaderIDDX9 *id);
+void ComputeFragmentShaderIDDX9(ShaderID *id);
 void GenerateFragmentShaderDX9(char *buffer);
 
 enum StencilValueType {
